api/requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
fastapi
pydantic
python-dotenv
uvicorn[standard]
psycopg2-binary


